Create a nice table from a summarised_result object.

Create a nice table from a summarised_result object.

Usage

formatSummarisedResult(
  result,
  tableType = "gt",
  header = NULL,
  estimateNameFormat = NULL,
  style = "default",
  decimals = c(integer = 0, numeric = 2, percentage = 2, proportion = 3),
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  .options = list()
)

Arguments

result

A summarised_result object.

tableType

Type of table to format into (currently accepting "gt").

header

Names of the columns to make headers. Names not corresponding to a column of the table result, will be used as headers at the defined position.

estimateNameFormat

Named list of estimate name's to join, sorted by computation order.

style

Named list that specifies how to style the different parts of the table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. If NULL, the table will not be styled (deafult of the chosed tableType), and if "deafult" it will be styled as OMOP's deafult.

decimals

Number of decimals per estimate_type (integer, numeric, percentage, proportion).

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

.options

list of deafult options (includeHeaderName = TRUE, keepNotFormatted = TRUE, decimalMark = ".", bigMark = ",", na = "-", groupNameCol = NULL, groupNameAsColumn = FALSE, groupOrder = NULL).

Value

A formatted table from a summarisedResult dataframe.

Examples

# \donttest{
mockSummarisedResult() |>
  formatSummarisedResult(tableType = "gt",
                         header = c("Study strata","strata_name",
                                    "strata_level"),
                         estimateNameFormat = c("N(%)" =
                                                  "<count> (<percentage>%)"),
                         style = "default",
                         decimals = c(
                           integer = 0, numeric = 2, percentage = 1,
                           proportion = 3
                         ),
                         .options = list("groupNameCol" = "group_level",
                                         "groupNameAsColumn" = FALSE,
                                         "groupOrder" = c("cohort1",
                                                          "cohort2"),
                                         "includeHeaderName" = FALSE))
#> <div id="orlvuzjcdm" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#orlvuzjcdm table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #orlvuzjcdm thead, #orlvuzjcdm tbody, #orlvuzjcdm tfoot, #orlvuzjcdm tr, #orlvuzjcdm td, #orlvuzjcdm th {
#>   border-style: none;
#> }
#> 
#> #orlvuzjcdm p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #orlvuzjcdm .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #orlvuzjcdm .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #orlvuzjcdm .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #orlvuzjcdm .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #orlvuzjcdm .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #orlvuzjcdm .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #orlvuzjcdm .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #orlvuzjcdm .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #orlvuzjcdm .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #orlvuzjcdm .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #orlvuzjcdm .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #orlvuzjcdm .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #orlvuzjcdm .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #orlvuzjcdm .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #orlvuzjcdm .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #orlvuzjcdm .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #orlvuzjcdm .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #orlvuzjcdm .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #orlvuzjcdm .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #orlvuzjcdm .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #orlvuzjcdm .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #orlvuzjcdm .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #orlvuzjcdm .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #orlvuzjcdm .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #orlvuzjcdm .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #orlvuzjcdm .gt_left {
#>   text-align: left;
#> }
#> 
#> #orlvuzjcdm .gt_center {
#>   text-align: center;
#> }
#> 
#> #orlvuzjcdm .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #orlvuzjcdm .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #orlvuzjcdm .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #orlvuzjcdm .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #orlvuzjcdm .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #orlvuzjcdm .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #orlvuzjcdm .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #orlvuzjcdm .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #orlvuzjcdm .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #orlvuzjcdm .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #orlvuzjcdm .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #orlvuzjcdm .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="11" style="background-color: #C8C8C8; font-weight: bold;" scope="colgroup" id></th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="9" style="background-color: #C8C8C8; font-weight: bold;" scope="colgroup" id="Study strata">
#>         <span class="gt_column_spanner">Study strata</span>
#>       </th>
#>     </tr>
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="cdm_name">cdm_name</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="result_type">result_type</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="package_name">package_name</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="package_version">package_version</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="group_name">group_name</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="variable_name">variable_name</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="variable_level">variable_level</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="estimate_name">estimate_name</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="estimate_type">estimate_type</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="additional_name">additional_name</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" style="font-weight: bold;" scope="col" id="additional_level">additional_level</th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="1" style="background-color: #E1E1E1; font-weight: bold;" scope="col" id="overall">
#>         <span class="gt_column_spanner">overall</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="4" style="background-color: #E1E1E1; font-weight: bold;" scope="colgroup" id="age_group and sex">
#>         <span class="gt_column_spanner">age_group and sex</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" style="background-color: #E1E1E1; font-weight: bold;" scope="colgroup" id="sex">
#>         <span class="gt_column_spanner">sex</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" style="background-color: #E1E1E1; font-weight: bold;" scope="colgroup" id="age_group">
#>         <span class="gt_column_spanner">age_group</span>
#>       </th>
#>     </tr>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="overall">overall</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="&amp;lt;40 and Male">&lt;40 and Male</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="&amp;gt;=40 and Male">&gt;=40 and Male</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="&amp;lt;40 and Female">&lt;40 and Female</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="&amp;gt;=40 and Female">&gt;=40 and Female</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="Male">Male</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="Female">Female</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="&amp;lt;40">&lt;40</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="font-weight: bold;" scope="col" id="&amp;gt;=40">&gt;=40</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <th colspan="20" class="gt_group_heading" scope="colgroup" id="cohort1">cohort1</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="cohort1  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort1  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort1  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort1  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort1  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort1  variable_name" class="gt_row gt_left">number subjects</td>
#> <td headers="cohort1  variable_level" class="gt_row gt_left">-</td>
#> <td headers="cohort1  estimate_name" class="gt_row gt_left">count</td>
#> <td headers="cohort1  estimate_type" class="gt_row gt_left">integer</td>
#> <td headers="cohort1  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right">9,970,691</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right">1,490,355</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right">5,185,566</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right">8,461,201</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right">7,182,697</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right">2,413,140</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right">5,470,434</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right">8,348,018</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right">  279,560</td></tr>
#>     <tr><td headers="cohort1  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort1  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort1  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort1  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort1  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort1  variable_name" class="gt_row gt_left">age</td>
#> <td headers="cohort1  variable_level" class="gt_row gt_left">-</td>
#> <td headers="cohort1  estimate_name" class="gt_row gt_left">mean</td>
#> <td headers="cohort1  estimate_type" class="gt_row gt_left">numeric</td>
#> <td headers="cohort1  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right">51.18</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right">83.56</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right">70.88</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right">87.42</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right"> 1.15</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right">88.82</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right">99.63</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right">50.02</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right">35.90</td></tr>
#>     <tr><td headers="cohort1  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort1  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort1  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort1  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort1  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort1  variable_name" class="gt_row gt_left">age</td>
#> <td headers="cohort1  variable_level" class="gt_row gt_left">-</td>
#> <td headers="cohort1  estimate_name" class="gt_row gt_left">sd</td>
#> <td headers="cohort1  estimate_type" class="gt_row gt_left">numeric</td>
#> <td headers="cohort1  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right"> 4.03</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right"> 7.70</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right"> 1.19</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right"> 1.95</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right"> 1.65</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right"> 6.63</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right"> 8.57</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right"> 9.27</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right"> 5.52</td></tr>
#>     <tr><td headers="cohort1  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort1  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort1  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort1  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort1  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort1  variable_name" class="gt_row gt_left">Medications</td>
#> <td headers="cohort1  variable_level" class="gt_row gt_left">Amoxiciline</td>
#> <td headers="cohort1  estimate_name" class="gt_row gt_left">N(%)</td>
#> <td headers="cohort1  estimate_type" class="gt_row gt_left">character</td>
#> <td headers="cohort1  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right">   90,944 (90.3%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right">    3,745 (58.1%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right">   59,355 (77.3%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right">   23,698 (99.5%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right">   90,630 (71.1%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right">   81,887 (21.5%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right">   69,983 (29.2%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right">   22,000 (72.2%)</td>
#> <td headers="cohort1  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right">   72,799 (86.7%)</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <th colspan="20" class="gt_group_heading" scope="colgroup" id="cohort2">cohort2</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="cohort2  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort2  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort2  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort2  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort2  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort2  variable_name" class="gt_row gt_left">number subjects</td>
#> <td headers="cohort2  variable_level" class="gt_row gt_left">-</td>
#> <td headers="cohort2  estimate_name" class="gt_row gt_left">count</td>
#> <td headers="cohort2  estimate_type" class="gt_row gt_left">integer</td>
#> <td headers="cohort2  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right">4,693,843</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right">8,056,800</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right">8,140,513</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right">4,039,110</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right">2,184,310</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right">4,183,614</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right">6,688,707</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right">5,076,503</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right">6,603,593</td></tr>
#>     <tr><td headers="cohort2  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort2  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort2  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort2  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort2  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort2  variable_name" class="gt_row gt_left">age</td>
#> <td headers="cohort2  variable_level" class="gt_row gt_left">-</td>
#> <td headers="cohort2  estimate_name" class="gt_row gt_left">mean</td>
#> <td headers="cohort2  estimate_type" class="gt_row gt_left">numeric</td>
#> <td headers="cohort2  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right">77.49</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right">58.45</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right">63.40</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right">85.87</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right">56.69</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right">25.30</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right">91.88</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right">86.74</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right">24.85</td></tr>
#>     <tr><td headers="cohort2  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort2  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort2  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort2  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort2  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort2  variable_name" class="gt_row gt_left">age</td>
#> <td headers="cohort2  variable_level" class="gt_row gt_left">-</td>
#> <td headers="cohort2  estimate_name" class="gt_row gt_left">sd</td>
#> <td headers="cohort2  estimate_type" class="gt_row gt_left">numeric</td>
#> <td headers="cohort2  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right"> 5.77</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right"> 6.87</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right"> 2.45</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right"> 0.45</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right"> 9.10</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right"> 0.71</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right"> 9.97</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right"> 6.12</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right"> 1.73</td></tr>
#>     <tr><td headers="cohort2  cdm_name" class="gt_row gt_left">mock</td>
#> <td headers="cohort2  result_type" class="gt_row gt_left">-</td>
#> <td headers="cohort2  package_name" class="gt_row gt_left">formatTable</td>
#> <td headers="cohort2  package_version" class="gt_row gt_right">0.0.1</td>
#> <td headers="cohort2  group_name" class="gt_row gt_left">cohort_name</td>
#> <td headers="cohort2  variable_name" class="gt_row gt_left">Medications</td>
#> <td headers="cohort2  variable_level" class="gt_row gt_left">Amoxiciline</td>
#> <td headers="cohort2  estimate_name" class="gt_row gt_left">N(%)</td>
#> <td headers="cohort2  estimate_type" class="gt_row gt_left">character</td>
#> <td headers="cohort2  additional_name" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  additional_level" class="gt_row gt_left">overall</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]overall
#> [column_name]overall" class="gt_row gt_right">   21,708 (23.8%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Male" class="gt_row gt_right">   45,623 ( 0.4%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Male" class="gt_row gt_right">   33,280 (94.4%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]<40 and Female" class="gt_row gt_right">   56,835 (43.8%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group and sex
#> [column_name]>=40 and Female" class="gt_row gt_right">   25,221 (75.1%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Male" class="gt_row gt_right">   46,401 (66.8%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]sex
#> [column_name]Female" class="gt_row gt_right">   91,766 (40.8%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]<40" class="gt_row gt_right">   97,284 (35.1%)</td>
#> <td headers="cohort2  [header]Study strata
#> [header_level]age_group
#> [column_name]>=40" class="gt_row gt_right">   81,908 (73.8%)</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>
# }