Creats a flextable from a dataframe using as deliminter (delim) for spanners.

Creats a flextable from a dataframe using as deliminter (delim) for spanners.

Usage

fxTable(
  x,
  delim = "\n",
  style = list(header = list(cell = officer::fp_cell(background.color = "#c8c8c8"), text
    = officer::fp_text(bold = TRUE)), header_name = list(cell =
    officer::fp_cell(background.color = "#d9d9d9"), text = officer::fp_text(bold =
    TRUE)), header_level = list(cell = officer::fp_cell(background.color = "#e1e1e1"),
    text = officer::fp_text(bold = TRUE)), column_name = list(text =
    officer::fp_text(bold = TRUE)), group_label = list(cell =
    officer::fp_cell(background.color = "#e9ecef", border = officer::fp_border(width = 1,
    
     color = "gray")), text = officer::fp_text(bold = TRUE))),
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupNameCol = NULL,
  groupNameAsColumn = FALSE,
  groupOrder = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body.

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupNameCol

Column to use as group labels.

groupNameAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupOrder

Order in which to display group labels.

Value

A flextable.

Examples

# \donttest{
mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  spanHeader(header = c("Study strata", "strata_level"),
             includeHeader = FALSE) |>
  fxTable(
    style = list(
      "header" = list(
        "cell" = officer::fp_cell(background.color = "#c8c8c8"),
        "text" = officer::fp_text(bold = TRUE)),
      "header_level" = list(
        "cell" = officer::fp_cell(background.color = "#e1e1e1"),
        "text" = officer::fp_text(bold = TRUE)),
      "group_label" = list(
        "cell" = officer::fp_cell(
          background.color = "#e9ecef",
          border = officer::fp_border(width = 1, color = "gray")),
        "text" = officer::fp_text(bold = TRUE))
    ),
    na = "--",
   title = "fxTable example",
   subtitle = NULL,
   caption = NULL,
  groupNameCol = "group_level",
   groupNameAsColumn = TRUE,
   groupOrder = c("cohort1", "cohort2")
 )

fxTable example

group_level

cdm_name

result_type

package_name

package_version

group_name

strata_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

<40 and Male

>=40 and Male

<40 and Female

>=40 and Female

Male

Female

<40

>=40

cohort1

mock

--

formatTable

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

9,970,691

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

1,490,355

5,185,566

8,461,201

7,182,697

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

2,413,140

5,470,434

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

8,348,018

  279,560

mock

--

formatTable

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

51.2

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

83.6

70.9

87.4

 1.1

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

88.8

99.6

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

50.0

35.9

mock

--

formatTable

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

 4.0

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

 7.7

 1.2

 1.9

 1.6

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

 6.6

 8.6

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

 9.3

 5.5

mock

--

formatTable

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

   90,944

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

    3,745

   59,355

   23,698

   90,630

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

   81,887

   69,983

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

   22,000

   72,799

mock

--

formatTable

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

90.2923798188567

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

58.1366044003516

77.3008481133729

99.5123026426882

71.097124973312

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

21.4942595688626

29.1757629951462

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

72.1759729320183

86.6615703329444

cohort2

mock

--

formatTable

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

4,693,843

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

8,056,800

8,140,513

4,039,110

2,184,310

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

4,183,614

6,688,707

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

5,076,503

6,603,593

mock

--

formatTable

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

77.5

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

58.4

63.4

85.9

56.7

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

25.3

91.9

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

86.7

24.9

mock

--

formatTable

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

 5.8

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

 6.9

 2.4

 0.4

 9.1

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

 0.7

10.0

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

 6.1

 1.7

mock

--

formatTable

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

   21,708

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

   45,623

   33,280

   56,835

   25,221

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

   46,401

   91,766

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

   97,284

   81,908

mock

--

formatTable

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

23.8453106489033

--

--

--

--

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

0.449630804359913

94.3516464205459

43.8137200428173

75.0603328226134

--

--

--

--

mock

--

formatTable

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

66.7815764434636

40.7973201479763

--

--

mock

--

formatTable

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

35.1248814957216

73.8091561477631

# }