Numeric vector with accounting format

Formats numbers with a thousand separator and two digits after the decimal point.

num_accounting(x, digits = 2L, format = "f", big.mark = ",", ...)

parse_accounting(
  x,
  digits = max(get_digits(x)),
  format = "f",
  big.mark = ",",
  ...
)

Arguments

x

a numeric vector.

digits

an integer to indicate the number of digits of the percentage string.

format

format type passed to formatC().

big.mark

thousands separator

...

additional parameters passed to formattable().

See also

Examples

num_accounting(15320)
#> [1] 15,320.00
num_accounting(-12500)
#> [1] (12,500.00)
num_accounting(c(1200, -3500, 2600), format = "d")
#> [1] 1,200   (3,500) 2,600  
parse_accounting(c("123,23.50", "(123.243)"))
#> [1] 12,323.500 (123.243)