Numeric vector with thousands separators

Formats numbers with a thousand separator and a prespecified number of digits after the decimal point.

num_comma(x, digits = 2L, format = "f", big.mark = ",", ...)

parse_comma(x, digits = max(get_digits(x)), format = "f", big.mark = ",", ...)

Arguments

x
a numeric vector.

digits
an integer to indicate the number of digits of the percentage string.

format
format type passed to formatC().

big.mark
thousands separator

...
additional parameters passed to formattable().

See also

Examples

num_comma(1000000)
#> [1] 1,000,000.00
num_comma(c(1250000, 225000))
#> [1] 1,250,000.00 225,000.00  
num_comma(c(1250000, 225000), format = "d")
#> [1] 1,250,000 225,000  
parse_comma("123,345.123")
#> [1] 123,345.123